AIF-043 M1: in-memory indexed tables (xbase::ramfs) - x64 table + native CDX-V64 in RAM, proven
In-process RAM VFS (xbase::ramfs) serving byte-identical DBF row images and a native CDX-V64 (RUN8/uint64) index built and traversed entirely in RAM; DO mem / VDISK MOUNT points DBF/INDEXES/LMDB slots under Slot::RAM. mem_proof.dts self-asserts (MEM_T1/T2/T3=.T., zero files on disk); registered as regression MEM.

Root-cause fix: src/core/dbf_create.cpp was globbed into the CLI exe and shadowed xbase.lib's ramfs-aware create_dbf at link (exe object wins over static-lib member), so CREATE wrote to disk while USE read RAM. Excluded from the CLI glob (record_view.cpp precedent) and parked to .bak. Ramfs-aware gates in cmd_cdx and SET ORDER (container + LMDB-env skip). Build hardened: configure-time duplicate-basename shadow guard + /INCREMENTAL:NO.

Docs: M1 proof closeout, RAM-DBF positioning/limitations, and teaching case ENG-060 (Native CDX vs LMDB-Backed CDX) with runtime proof.

+++ include/xindex/inx_payload.hpp
@@ -11,7 +11,10 @@ namespace xindex {
 
 struct InxEntry {
     std::string key;
-    std::uint32_t recno{0};
+    // RECNO64 "build for big": recno is uint64 in memory. Legacy 1INX/2INX and
+    // CNX RUN1 payloads store 4-byte recnos and read as the fallback (widened on
+    // load); the wide CDX RUN format stores 8-byte recnos.
+    std::uint64_t recno{0};
 };
 
 class InxPayload final {
@@ -40,7 +43,7 @@ public:
     const InxEntry& entryAt(std::size_t i) const;
 
     bool hasPosByRecno() const noexcept { return !pos_by_recno_.empty(); }
-    std::int32_t positionOfRecno(std::uint32_t recno) const noexcept;
+    std::int32_t positionOfRecno(std::uint64_t recno) const noexcept;
 
     // payload-level lookup helpers
     std::optional<std::size_t> topPos() const noexcept;
